refactor(test): relocate two nemoclaw-start blocks to trim file below legacy budget
test/nemoclaw-start.test.ts had grown to 5043 lines, 96 lines above its
legacy budget of 4947. Move the self-contained "gateway preload process
detection" and "persistent gateway log hardening" describe blocks to a new
test/nemoclaw-start-preload-gateway.test.ts with their required helper
functions duplicated inline. Original file now sits at 4937 lines; budget
entry updated accordingly.
