+++ ci/test-file-size-budget.json
@@ -8,7 +8,7 @@
     "test/channels-add-preset.test.ts": 1871,
     "test/generate-openclaw-config.test.ts": 1982,
     "test/install-preflight.test.ts": 4006,
-    "test/nemoclaw-start.test.ts": 5043,
+    "test/nemoclaw-start.test.ts": 4937,
     "test/onboard-messaging.test.ts": 2062,
     "test/onboard-selection.test.ts": 6867,
     "test/onboard.test.ts": 4774,
